Add more flutter build ios-framework tests before the impending jonahpocalypse
