+++ dev/devicelab/lib/framework/apk_utils.dart
@@ -76,36 +76,6 @@ Future<Iterable<String>> getFilesInAar(String aar) {
   return getFilesInApk(aar);
 }
 
-void checkItContains<T>(Iterable<T> values, Iterable<T> collection) {
-  for (T value in values) {
-    if (!collection.contains(value)) {
-      throw TaskResult.failure('Expected to find `$value` in `${collection.toString()}`.');
-    }
-  }
-}
-
-void checkItDoesNotContain<T>(Iterable<T> values, Iterable<T> collection) {
-  for (T value in values) {
-    if (collection.contains(value)) {
-      throw TaskResult.failure('Did not expect to find `$value` in `$collection`.');
-    }
-  }
-}
-
-///  Checks that [str] contains the specified [Pattern]s, otherwise throws
-/// a [TaskResult].
-void checkFileContains(List<Pattern> patterns, String filePath) {
-  final String fileContent = File(filePath).readAsStringSync();
-  for (Pattern pattern in patterns) {
-    if (!fileContent.contains(pattern)) {
-      throw TaskResult.failure(
-        'Expected to find `$pattern` in `$filePath` '
-        'instead it found:\n$fileContent'
-      );
-    }
-  }
-}
-
 TaskResult failure(String message, ProcessResult result) {
   print('Unexpected process result:');
   print('Exit code: ${result.exitCode}');
